HSEARCH-2223 Move tests packages from 'solr' to 'definition'
